Refactor `relvar` facade and de-bloat `Database` module
- Replaced glob import `pub use relvar_core::*;` in `relvar/src/lib.rs` with explicit exports to enforce a clean public API.
- Introduced `relvar/src/prelude.rs` for user convenience.
- Extracted ~1100 lines of tests from `relvar-core/src/database/mod.rs` to `relvar-core/src/database/tests.rs` to reduce module bloat.
- Moved `relvar-core/src/query.rs` to `relvar-core/src/query/mod.rs` to prepare for future expansion.
- Updated integration tests and documentation examples to align with the new API structure.
