+++ relvar/src/prelude.rs
@@ -1,0 +1,9 @@
+//! The Relvar Prelude
+//!
+//! Re-exports commonly used types for convenience.
+
+pub use relvar_core::database::Database;
+pub use relvar_core::values::{Relation, Tuple, ScalarValue};
+pub use relvar_core::types::{RelationType, TupleType, ScalarType};
+pub use relvar_core::query::Query;
+pub use relvar_core::storage_engine::InMemoryEngine;
